requirements.txt
django
djangorestframework
psycopg2-binary
python-dotenv
pylint
black
httpie